fix(Snap): Round off decimal into 2 points

+++ apps/snap/serializers/price_monitor_serializers.py
@@ -60,7 +60,7 @@ class OverviewPriceMonitorSnapReport(serializers.Serializer):
     min = serializers.FloatField(source='min_value')
     max = serializers.FloatField(source='max_value')
     mode = serializers.FloatField(source='mode_value')
-    mean = serializers.DecimalField(max_digits=10, decimal_places=1, source='mean_value')
+    mean = serializers.DecimalField(max_digits=10, decimal_places=2, source='mean_value')
 
 
 class MonthPriceMonitorSnapReportSerializer(serializers.Serializer):
@@ -74,13 +74,13 @@ class BrandoverviewPriceMonitorSnapReportSerializer(serializers.Serializer):
     min = serializers.FloatField(source='min_value')
     max = serializers.FloatField(source='max_value')
     mode = serializers.FloatField(source='mode_value')
-    mean = serializers.DecimalField(max_digits=10, decimal_places=1, source='mean_value')
+    mean = serializers.DecimalField(max_digits=10, decimal_places=2, source='mean_value')
 
 
 class CountryPriceMonitorSnapReportSerializer(serializers.Serializer):
     country = serializers.CharField(source='country_name')
     sku = serializers.CharField(source='sku_name')
-    value = serializers.FloatField()
+    value = serializers.DecimalField(max_digits=10, decimal_places=2)
 
 
 class VisitPerCityPriceMonitorSnapReportSerializer(AnswerPerCityReportSerializer):
